+++ Package.swift
@@ -1,5 +1,10 @@
+// swift-tools-version:5.0
 import PackageDescription
 
 let package = Package(
-    name: "SwiftHook"
+    name: "SwiftHook",
+    targets: [
+        .target(name: "SwiftHook"),
+        .testTarget(name: "SwiftHookTests", dependencies: ["SwiftHook"]),
+    ]
 )